recebimento de muudanças para adicionar(ok)

+++ src/main/java/br/com/dao/CursosDao.java
@@ -3,9 +3,10 @@ package br.com.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 import br.com.curso.controller.javabeans.Cursos;
 
 @Repository
- public interface CursosDao extends JpaRepository<Cursos, Integer>{
+ public interface CursosDao extends JpaRepository<Cursos, Long>{
 	
 }
